requirements.txt
# 基础依赖
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# 数据库
mysql-connector-python==8.2.0
SQLAlchemy==2.0.25
PyMySQL==1.1.0

# API和网络
requests==2.31.0
aiohttp==3.9.1
urllib3==2.1.0

# 测试
pytest==7.4.4
pytest-cov==4.1.0
pytest-qt==4.3.1
pytest-asyncio==0.23.3
pytest-xdist==3.5.0

# 工具和实用程序
python-dotenv==1.0.0
pydantic==2.5.3
typing-extensions==4.9.0
loguru==0.7.2

# 性能监控
psutil==5.9.7
memory-profiler==0.61.0

# 开发工具
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# 文档
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# 安全
cryptography==41.0.7
bcrypt==4.1.2

# 缓存
redis==5.0.1
cachetools==5.3.2

# 其他依赖
beautifulsoup4==4.12.3
fake-useragent==1.4.0
pandas==2.2.0
lxml==5.1.0
python-dateutil==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
six==1.16.0
openai>=1.0.0
matplotlib>=3.7.0
tk>=0.1.0
transformers>=4.11.0
torch>=1.10.0
numpy>=1.21.0
scikit-learn>=0.24.0
tensorflow>=2.7.0
accelerate>=0.27.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
einops>=0.7.0
protobuf>=4.25.0 